requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.5
gunicorn==23.0.0
Werkzeug==3.0.4
WTForms==3.1.2
